pyproject.toml
[project]
name = "spark-graphframes-example"
version = "0.1.0"
description = "Sample Spark + GraphFrames application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphframes-py>=0.9.3",
    "pyspark>=4.0.1",
]
